+++ containup/cli.py
@@ -20,7 +20,6 @@ class Config:
 
     def __init__(self, args: argparse.Namespace):
         self._args: argparse.Namespace = args
-        print(args.service)
 
     @property
     def extra_args(self) -> List[str]:
@@ -29,7 +28,13 @@ class Config:
         Returns:
             List of strings or empty list
         """
-        return self._args.extra_args or []
+        extra_args: List[str] = self._args.extra_args
+        extra: List[str] = (
+            (extra_args or [])[1:]
+            if (extra_args or [])[:1] == ["--"]
+            else (extra_args or [])
+        )
+        return extra
 
     @property
     def command(self) -> str:
@@ -51,7 +56,7 @@ class Config:
     @property
     def services(self) -> list[str]:
         """Returns service as list or empty list"""
-        return self._args.service
+        return self._args.service or []
 
     def validate(self) -> None:
         """Validates configuration and stops programm if there are errors"""
@@ -81,30 +86,61 @@ def containup_cli_args(prog: str, known_args: list[str]) -> Config:
     )
     subparsers = parser.add_subparsers(dest="command", required=True)
 
-    p = subparsers.add_parser("up")
-    p.add_argument(
-        "service", nargs="*", help="If specified, launches only those services"
+    up_parser = subparsers.add_parser("up")
+    up_parser.add_argument(
+        "--service",
+        nargs="*",
+        help="If specified, launches only those services",
     )
-    p.add_argument("extra_args", nargs=argparse.REMAINDER, help="Your own arguments")
+    _add_extra_args(up_parser)
 
-    p = subparsers.add_parser("down")
-    p.add_argument("service", nargs="*", help="If specified, stops only those services")
-    p.add_argument("extra_args", nargs=argparse.REMAINDER, help="Your own arguments")
+    down_parser = subparsers.add_parser("down")
+    down_parser.add_argument(
+        "--service", nargs="*", help="If specified, stops only those services"
+    )
+    _add_extra_args(down_parser)
 
     logs_parser = subparsers.add_parser("logs")
-    logs_parser.add_argument("service", nargs=1, help="Get logs from service")
-    p.add_argument("extra_args", nargs=argparse.REMAINDER, help="Your own arguments")
+    logs_parser.add_argument("--service", help="Get logs from service", required=True)
+    _add_extra_args(logs_parser)
 
-    subparsers.add_parser("export")
-    p.add_argument("extra_args", nargs=argparse.REMAINDER, help="Your own arguments")
+    export_parser = subparsers.add_parser("export")
+    _add_extra_args(export_parser)
 
     args = parser.parse_args(args=known_args)
-
     config = Config(args)
     config.validate()
+
+    print("Namespace: ", args)
 
     return config
 
 
+def _add_extra_args(parser: argparse.ArgumentParser) -> None:
+    parser.add_argument(
+        "extra_args", nargs=argparse.REMAINDER, help="Your own arguments"
+    )
+
+
 if __name__ == "__main__":
-    print(containup_cli())
+    #
+    # Tool to debug command line parsing
+    #
+    # Just an example of reading the CLI and extracting additional
+    # parameters
+    #
+
+    # call our CLI parser
+    containup_config = containup_cli()
+
+    # get your extra arguments (it's a list of string like sys.argv[:1])
+    script_extra_args = containup_config.extra_args
+
+    # Then, you can parse them with, for example, Python's argparse :
+    sub_parser = argparse.ArgumentParser(prog="mystack")
+    sub_parser.add_argument("--origin_name")
+    sub_parser.add_argument("--origin_version")
+
+    parsed = sub_parser.parse_args(args=script_extra_args)
+    print("origin_name=", parsed.origin_name)
+    print("origin_name=", parsed.origin_version)